refactor: move i18n config and translations to a separate folder
