+++ src/i18n/i18n-config.js
@@ -1,0 +1,5 @@
+export const I18N_CONFIG = {
+  disableRedirectToSelectedLanguage: true,
+  hideEsFromAlgoliaSearch: true,
+  hideLanguageToggle: true,
+};
