--- src/utils/i18n-config.js
@@ -1,5 +1,0 @@
-export const I18N_CONFIG = {
-  disableRedirectToSelectedLanguage: false,
-  hideEsFromAlgoliaSearch: true,
-  hideLanguageToggle: false,
-};
